+++ prototype/app/src/main/java/com/example/make_a_story_prototype/main/Characters/view/ImageCards/CharacterImageCardHolder.java
@@ -1,7 +1,12 @@
-package com.example.make_a_story_prototype.main.Characters.view;
+package com.example.make_a_story_prototype.main.Characters.view.ImageCards;
 
+import android.graphics.PorterDuff;
+import android.graphics.PorterDuffColorFilter;
 import android.graphics.drawable.Drawable;
 import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
+import android.widget.ImageView;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 
@@ -13,21 +18,24 @@ import androidx.annotation.NonNull;
 import androidx.cardview.widget.CardView;
 import androidx.recyclerview.widget.RecyclerView;
 
-public class CharacterNameCardHolder extends RecyclerView.ViewHolder{
-    private TextView characterName;
+
+public class CharacterImageCardHolder extends RecyclerView.ViewHolder{
+
+    private ImageView characterImage;
     private RelativeLayout parentLayout;
     private CharacterCardItemViewModel vm;
 
-    public CharacterNameCardHolder(@NonNull View itemView) {
+    public CharacterImageCardHolder(@NonNull View itemView) {
         super(itemView);
-        this.characterName = itemView.findViewById(R.id.characterName);
+        this.characterImage = itemView.findViewById(R.id.characterImage);
         this.parentLayout = itemView.findViewById(R.id.parentLayout);
     }
 
     public void setViewModel(CharacterCardItemViewModel vm) {
         this.vm = vm;
 
-        characterName.setText(vm.cardItem.getImageLabel());
+        // null?
+        characterImage.setImageResource(vm.cardItem.getImageResource());
 
         Drawable imageBackground = parentLayout.getBackground();
         Util.changeDrawableColor(imageBackground, vm.backgroundColor);
